Anwendungen werden zufällig über alle verfügbaren Controller verteilt.
git-svn-id: svn://localhost/twospot_svn/trunk@799 8d73be03-56cb-46cc-8948-98b08ae92ef2

+++ Controller/src/org/prot/controller/management/jmx/IJmxResources.java
@@ -10,7 +10,7 @@ public interface IJmxResources
 
 	public long freeMemory();
 
-	public long requestsPerSecond();
+	public double requestsPerSecond();
 
 	public Set<String> getApps();
